Fixed a chaining bug, fixed 2 cmdline bugs, misc other.
* Bugfix1: CFUPSBase::__wakeup() was declared protected instead
of public. This was causing chaining
* Bugfix2: an absolute filepath for the settings file on the
commandline (the -i option to fups.php) caused bad cookie
and serialisation filenames to be generated. This required
fixes to make_cookie_filename() and make_serializefilename()
in common.php.
* Bugfix3: $redirect was set to empty instead of ' 1>/dev/null'
in make_php_exec_cmd() for commandline invocations - this
caused the script to fail to exit after calling exit after
calling exec() in CFUPSBase::check_do_chain().
* Bugfixes4: prepended $this-> to a couple of instances of the
write_err() method.
* Ensured that the cookie file is saved when chaining, and
used this guarantee to avoid re-logging in.
* Wrapped commandline arguments in escapeshellarg() just to
be sure.
* Improved debug messages.