Fixes for running under IIS under Windows.
The biggest fix was to change the hiving-off of the background PHP
process from exec() to popen(), because there seems to be no way under
Windows+IIS to prevent PHP from waiting for an exec()ed process to
finish before continuing. The fix is to use popen() and prepend the
command with "start /b ", which is dangerous because if the command is
not runnable, "start" tries to pop up an error dialog, which in web-
server environment is invisible but causes a hang until PHP times out.
There is probably no guaranteed way to avoid this, but my current best
attempt is the script would_run.bat, which checks that at least the
command is runnable given the current directory and path - probably this
will still cause problems if the process doesn't have permission to
execute the command anyway, but I haven't tested that much.