Added support for rebasing img/anchor URLs. Misc other.
* Relative image and anchor URLs in posts are now converted into
the correct absolute URLs, so images should now always display
(assuming an internet connection) and links in posts should
now always direct to the correct place.
* Bugfix: post counts were sometimes doubled on phpBB forums due
to the similarity of the prosilver.1 and prosilver.2
'search_results_page_data' regexes. Combining these into a
single regex fixed the problem.
* Updated the answer to the "Which skins are supported?"
question in the XenForo Q&A.