Improved post content detection for phpBB's prosilver skin.
* Improved a couple of phpBB prosilver_3.1.6 regexes:
* 'post_contents_ext' to allow for detection of posts with attachment
boxes (though this does not resolve the limitation by which attachments
are not downloaded when scraping by forums, i.e., when filling in the
"Forum IDs" setting via the web interface or when setting "forum_ids"
via the commandline interface).
* 'post_contents' to allow for attachments with a class of "file" on
top of "thumbnail", and to allow for posts with signatures or that have
been edited.
* Removed the 'post_contents' regex added on 2018-10-25 in commit
5e4febfcf507027c1f9e76ec4e6a4fd9be4e3ca1, which is not only unnecessary
given these changes but incomplete and thus sometimes gives incorrect
results.
* Reordered the phpBB prosilver regexes so that the most recent are
topmost.