Added double-asterisked fields on the option entry pages.
Double-asterisked fields indicate that at least one of these fields is
required.

This is a badly needed change as a significant proportion of FUPS error
emails that I receive is due to people not filling out one of these
fields, which until now had had no indication whatsoever as to the
necessity of at least one of them being filled out.